画像をリサイズして登録するコンポーネントを作成
