+++ src/components/container/templates/users/sign_up.js
@@ -12,14 +12,6 @@ import TextareaAutosize from "@material-ui/core/TextareaAutosize";
 import EditUserIcon from "Molecules/edit_user_icon";
 import getIconURL from "Modules/getIconURL";
 import env from "env";
-import Cropper from "react-easy-crop";
-import Slider from "@material-ui/core/Slider";
-import { Button } from "@material-ui/core";
-import UserIcon from "Atoms/user_icon";
-import getCroppedImgData from "Modules/getCroppedImgData";
-import convertFileToDataURL from "Modules/convertFileToDataURL";
-import styles from "./sign_up.css";
-import dataURLtoFile from "Modules/dataURLtoFile";
 
 const ROOT_URL = env.ROOT_URL;
 
@@ -75,7 +67,7 @@ class SignUp extends Component {
     values.profile = this.state.textbox;
 
     // ユーザのアイコンをセット
-    let iconImage = this.state.croppedFile;
+    let iconImage = this.refs.EditUserIcon.getUserIcon();
 
     // 登録
     await this.props.postUserEvent(values, iconImage);
@@ -156,63 +148,6 @@ class SignUp extends Component {
     );
   }
 
-  onCropChange = (crop) => {
-    this.setState({ crop });
-  };
-
-  onCropComplete = (croppedArea, croppedAreaPixels) => {
-    this.setState({
-      croppedAreaPixels,
-    });
-  };
-
-  onZoomChange = (zoom) => {
-    this.setState({ zoom });
-  };
-
-  onFileChange = async (e) => {
-    if (e.target.files && e.target.files.length > 0) {
-      const file = e.target.files[0];
-      let imageDataUrl = await convertFileToDataURL(file);
-
-      this.setState({
-        imageSrc: imageDataUrl,
-        crop: { x: 0, y: 0 },
-        zoom: 1,
-      });
-    }
-  };
-
-  showResult = async () => {
-    try {
-      this.setState({
-        isCropping: true,
-      });
-
-      const resultArr = await getCroppedImgData(
-        this.state.imageSrc,
-        this.state.croppedAreaPixels
-      );
-      const croppedImageURL = resultArr[0];
-      const croppedImageDataURL = resultArr[1];
-
-      const croppedFile = dataURLtoFile(croppedImageDataURL);
-
-      this.setState({
-        croppedImage: croppedImageURL,
-        croppedFile: croppedFile,
-        isCropping: false,
-      });
-
-      // リサイズで表示する画像を消す
-      this.setState({ imageSrc: null });
-    } catch (e) {
-      this.setState({
-        isCropping: false,
-      });
-    }
-  };
-
   // プロフィール
   handleChange(e) {
     if (e.target.value.length > 1000) {
@@ -238,152 +173,85 @@ class SignUp extends Component {
       // 初期表示トピック
       const initTopics = "";
 
-      var Icon;
-      if (this.state.croppedImage) {
-        Icon = (
-          <div>
-            <UserIcon iconData={this.state.croppedImage} />
-          </div>
-        );
-      }
       return (
         <React.Fragment>
           <form onSubmit={handleSubmit(this.onSubmit)}>
-            <div className={styles.test}>ユーザ登録</div>
-            {Icon}
-            <div className={styles.App}>
-              <input
-                type="file"
-                accept="image/*"
-                onChange={this.onFileChange}
-              />
-              {this.state.imageSrc && (
-                <div>
-                  <div className={styles.cropContainer}>
-                    <Cropper
-                      image={this.state.imageSrc}
-                      crop={this.state.crop}
-                      zoom={this.state.zoom}
-                      aspect={this.state.aspect}
-                      onCropChange={this.onCropChange}
-                      onCropComplete={this.onCropComplete}
-                      onZoomChange={this.onZoomChange}
-                    />
-                  </div>
-                  <div className={styles.controls}>
-                    <Slider
-                      value={this.state.zoom}
-                      min={1}
-                      max={3}
-                      step={0.1}
-                      aria-labelledby="Zoom"
-                      onChange={(e, zoom) => this.onZoomChange(zoom)}
-                      // classes={{ container: "slider" }}
-                    />
-                  </div>
-                  <div className={styles.button}>
-                    <Button
-                      color="primary"
-                      variant="contained"
-                      onClick={this.showResult}
-                      disabled={this.state.isCropping}
-                    >
-                      決定
-                    </Button>
-                  </div>
-                  {/* <ImgDialog
-                  img={this.state.croppedImage}
-                  onClose={this.onClose}
-                /> */}
-                </div>
-              )}
-              {/* </div> App */}
-              {/* <EditUserIcon
+            <div>ユーザ登録</div>
+            <EditUserIcon
               defaultIconURL=""
               icon={this.state.defaultIconURL}
               ref="EditUserIcon"
-            /> */}
-
+            />
+            ユーザ名:
+            <Field
+              label="2文字以上20文字以内"
+              name="user_name"
+              disabled={submitting}
+              type="text"
+              component={this.renderField}
+              onBlur={(e) => this.onBlurUserName(e)}
+            />
+            <div>
+              メールアドレス:
+              <Field
+                label="メールアドレス"
+                name="email"
+                type="text"
+                component={this.renderField}
+                onBlur={(e) => this.onBlurEmail(e)}
+              />
+            </div>
+            <div>
+              興味のあるトピック
+              <TopicSelectBox
+                allTopics={allTopics}
+                initTopics={initTopics}
+                ref="TopicSelectBox"
+              />
+            </div>
+            <div>
+              プロフィール
               <div>
-                ユーザ名:
-                <Field
-                  label="2文字以上20文字以内"
-                  name="user_name"
-                  disabled={submitting}
-                  type="text"
-                  component={this.renderField}
-                  onBlur={(e) => this.onBlurUserName(e)}
+                <TextareaAutosize
+                  aria-label="profile"
+                  rowsMin={3}
+                  rowsMax={20}
+                  placeholder="1000文字以内"
+                  onChange={this.handleChange}
                 />
               </div>
-
-              <div>
-                メールアドレス:
-                <Field
-                  label="メールアドレス"
-                  name="email"
-                  type="text"
-                  component={this.renderField}
-                  onBlur={(e) => this.onBlurEmail(e)}
-                />
-              </div>
-
-              <div>
-                興味のあるトピック
-                <TopicSelectBox
-                  allTopics={allTopics}
-                  initTopics={initTopics}
-                  ref="TopicSelectBox"
-                />
-              </div>
-
-              <div>
-                プロフィール
-                <div>
-                  <TextareaAutosize
-                    aria-label="profile"
-                    rowsMin={3}
-                    rowsMax={20}
-                    placeholder="1000文字以内"
-                    onChange={this.handleChange}
-                  />
-                </div>
-              </div>
-
-              <div>
-                パスワード
-                <Field
-                  label="パスワード"
-                  name="password"
-                  type="password"
-                  component={this.renderField}
-                />
-              </div>
-
-              <div>
-                確認用パスワード
-                <Field
-                  label="確認用パスワード"
-                  name="confirm_password"
-                  type="password"
-                  component={this.renderField}
-                />
-              </div>
-
-              <div>
-                <input
-                  type="submit"
-                  value="Submit"
-                  disabled={
-                    isError ||
-                    isDuplicateName ||
-                    isDuplicateEmail ||
-                    submitting ||
-                    this.state.textBoxFlag
-                  }
-                />
-              </div>
-            </div>{" "}
-            {/* App */}
+            </div>
+            <div>
+              パスワード
+              <Field
+                label="パスワード"
+                name="password"
+                type="password"
+                component={this.renderField}
+              />
+            </div>
+            <div>
+              確認用パスワード
+              <Field
+                label="確認用パスワード"
+                name="confirm_password"
+                type="password"
+                component={this.renderField}
+              />
+            </div>
+            <div>
+              <input
+                type="submit"
+                value="Submit"
+                disabled={
+                  isError ||
+                  isDuplicateName ||
+                  isDuplicateEmail ||
+                  submitting ||
+                  this.state.textBoxFlag
+                }
+              />
+            </div>
           </form>
           <div>
             <Link to={`/login`}>ログイン</Link>
