+++ src/components/presentational/molecules/edit_user_icon.css
@@ -1,10 +1,10 @@
-.App {
+/* .App {
   position: absolute;
   top: 0;
   left: 0;
   right: 0;
   bottom: 0;
-}
+} */
 
 .cropContainer {
   width: 100%;
